🔧 fix(extensions): use .ts extensions in internal imports
